Split LSPS1 in client/server handler modules
